plogging 활동 종료 api구현

+++ app/src/main/java/com/example/planet/data/dto/PloggingInfo.kt
@@ -7,7 +7,7 @@ data class PloggingInfo(
     val trashCount: List<Map<String, Int>>,
     val distance: Double,
     val kcal: Double,
-    val speed : Double,
+    val speed : Long,
     val score: Int,
     val ploggingTime: Long
 )
